prosia_extensions

Module arrays

Enums

ExtremaError

Error type for extrema operations

Traits

ArrayExtrema

Extension trait providing convenience methods for computing extrema (minima, maxima) and their indices on ndarray::ArrayBase values.

Integrable

Sequence

Trait providing monotonicity checks for 1-D ndarray arrays.

Functions

find_index_ge

Returns the index of the smallest element in a sorted slice that is greater than or equal to the given value.

find_index_le

Returns the index of the largest element in a sorted slice that is less than or equal to the given value.

lower_bound_index

Returns the index of the first element in a sorted slice that is greater than or equal to val.

upper_bound_index

Returns the index of the last element ≤ val in a sorted slice.